refactor magic method to module
